# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/auth/me": {
      "get": {
        "operationId": "AuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "auth"
        ]
      },
      "patch": {
        "operationId": "AuthController_update",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAccountDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/change-password": {
      "post": {
        "operationId": "AuthController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/businesses": {
      "post": {
        "operationId": "BusinessesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBusinessDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "businesses"
        ]
      }
    },
    "/api/v1/businesses/{id}": {
      "get": {
        "operationId": "BusinessesController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "businesses"
        ]
      },
      "patch": {
        "operationId": "BusinessesController_patch",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBusinessDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "businesses"
        ]
      }
    },
    "/api/v1/businesses/{id}/branches": {
      "post": {
        "operationId": "BusinessesController_branch",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBranchDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "businesses"
        ]
      },
      "get": {
        "operationId": "BusinessesController_branches",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "businesses"
        ]
      }
    },
    "/api/v1/businesses/{id}/branches/{branchId}": {
      "patch": {
        "operationId": "BusinessesController_updateBranch",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "branchId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBranchDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "businesses"
        ]
      }
    },
...
    "/api/v1/admin/drivers/{id}": {
      "get": {
        "operationId": "AdminController_driverDetail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/wallets": {
      "get": {
        "operationId": "AdminController_wallets",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/wallets/{id}": {
      "get": {
        "operationId": "AdminController_walletDetail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/drivers/{id}/review": {
      "post": {
        "operationId": "AdminController_review",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveDriverDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/whatsapp-identities": {
      "post": {
        "operationId": "AdminController_identity",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminWhatsAppIdentityDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
...
    "/api/v1/admin/finance/adjustments": {
      "post": {
        "operationId": "AdminController_adjustment",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdjustmentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/service-zones": {
      "post": {
        "operationId": "AdminController_zone",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ZoneDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      },
      "get": {
        "operationId": "AdminController_zonesList",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/matching/run": {
      "post": {
        "operationId": "AdminController_match",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/matching-config": {
      "get": {
        "operationId": "AdminController_matchingConfig",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/pricing-rules": {
      "get": {
        "operationId": "AdminController_pricing",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/{resource}": {
      "get": {
        "operationId": "AdminController_list",
        "parameters": [
          {
            "name": "resource",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    }
  },
  "info": {
    "title": "ORBITA",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "IdentityDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "role": {
            "type": "string",
            "enum": [
              "ADMIN",
              "BUSINESS_OWNER",
              "BUSINESS_STAFF",
              "DRIVER"
            ]
          },
          "businessId": {
            "type": "string",
            "nullable": true,
            "format": "uuid"
          },
          "driverId": {
            "type": "string",
            "nullable": true,
            "format": "uuid"
          },
          "phone": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "name",
          "email",
          "role",
          "businessId",
          "driverId",
          "phone"
        ]
      },
      "UpdateAccountDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          }
        }
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string"
          },
          "newPassword": {
            "type": "string"
          }
        },
        "required": [
          "currentPassword",
          "newPassword"
        ]
      },
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "BUSINESS_OWNER",
              "DRIVER"
            ]
          },
          "phone": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password",
          "name",
          "role",
          "phone"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "RefreshDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "CreateBusinessDto": {
        "type": "object",
        "properties": {
          "legalName": {
            "type": "string"
          },
          "tradeName": {
            "type": "string"
          },
          "document": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "category": {
            "type": "string",
            "enum": [
              "FOOD",
              "PHARMACY",
              "BEVERAGE_DISTRIBUTOR"
            ]
          }
        },
        "required": [
          "legalName",
          "tradeName",
          "document",
          "phone",
          "email",
          "category"
        ]
      },
      "UpdateBusinessDto": {
        "type": "object",
        "properties": {
          "legalName": {
            "type": "string"
          },
          "tradeName": {
            "type": "string"
          },
          "document": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "category": {
            "type": "string",
            "enum": [
              "FOOD",
              "PHARMACY",
              "BEVERAGE_DISTRIBUTOR"
            ]
          }
        }
      },
      "CreateBranchDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "latitude": {
            "type": "number"
          },
          "longitude": {
            "type": "number"
          },
          "timezone": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "address",
          "latitude",
          "longitude"
        ]
      },
      "UpdateBranchDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "latitude": {
            "type": "number"
          },
          "longitude": {
            "type": "number"
          },
          "timezone": {
            "type": "string"
          },
          "active": {
            "type": "boolean"
          }
        }
      },
...
      "ApproveDriverDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "APPROVED",
              "REJECTED"
            ]
          }
        },
        "required": [
          "status"
        ]
      },
      "AdminWhatsAppIdentityDto": {
        "type": "object",
        "properties": {
          "phoneNumber": {
            "type": "string"
          },
          "entityType": {
            "type": "string",
            "enum": [
              "BUSINESS",
              "DRIVER",
              "CUSTOMER"
            ]
          },
          "entityId": {
            "type": "string"
          },
          "evidence": {
            "type": "string",
            "description": "Reference to recorded verification and opt-in evidence"
          }
        },
        "required": [
          "phoneNumber",
          "entityType",
          "entityId",
          "evidence"
        ]
      },
      "AdjustmentDto": {
        "type": "object",
        "properties": {
          "walletId": {
            "type": "string"
          },
          "amountCents": {
            "type": "number"
          },
          "reason": {
            "type": "string"
          },
          "idempotencyKey": {
            "type": "string"
          }
        },
        "required": [
          "walletId",
          "amountCents",
          "reason",
          "idempotencyKey"
        ]
      },
      "ZoneDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "boundary": {
            "type": "object"
          }
        },
        "required": [
          "name",
          "city",
          "boundary"
        ]
      }
    }
  }
}